+++ CHANGELOG.md
@@ -11,7 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Open Images format now supports bounding box annotations
   (<https://github.com/openvinotoolkit/datumaro/pull/352>).
 - Bounding boxes values decrement transform (<https://github.com/openvinotoolkit/datumaro/pull/366>)
-
 ### Changed
 - TBD
 
@@ -22,7 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TBD
 
 ### Fixed
-- Patching of datasets in Datumaro, CVAT, COCO and CIFAR formats (<https://github.com/openvinotoolkit/datumaro/pull/365>, <https://github.com/openvinotoolkit/datumaro/pull/347>, <https://github.com/openvinotoolkit/datumaro/pull/346>)
+- Patching of datasets in Datumaro, CVAT, COCO, CIFAR and Open Images formats
+  (<https://github.com/openvinotoolkit/datumaro/pull/365>,
+  <https://github.com/openvinotoolkit/datumaro/pull/347>,
+  <https://github.com/openvinotoolkit/datumaro/pull/346>,
+  <https://github.com/openvinotoolkit/datumaro/pull/363>)
+  <https://github.com/openvinotoolkit/datumaro/pull/346>,
+  <https://github.com/openvinotoolkit/datumaro/pull/363>)
 - Unsafe unpickling in CIFAR import (<https://github.com/openvinotoolkit/datumaro/pull/362>)
 
 ### Security